# swagger.yaml
openapi: 3.0.0
info:
  title: Mini Wallet API
  description: API for managing a mini wallet application with user authentication, wallet operations, and transaction management
  version: 1.0.0
  contact:
    email: support@miniwallet.com
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.miniwallet.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    FundRequest:
      type: object
      required:
        - walletId
        - amount
      properties:
        walletId:
          type: string
          format: uuid
          description: ID of the wallet to fund
        amount:
          type: number
          format: float
          minimum: 0.01
          description: Amount to fund the wallet with
    TransferRequest:
      type: object
      required:
        - senderWalletId
        - receiverWalletId
        - amount
      properties:
        senderWalletId:
          type: string
          format: uuid
          description: ID of the sender's wallet
        receiverWalletId:
          type: string
          format: uuid
          description: ID of the receiver's wallet
        amount:
          type: number
          format: float
          minimum: 0.01
          description: Amount to transfer
    WithdrawRequest:
      type: object
      required:
        - walletId
        - amount
      properties:
        walletId:
          type: string
          format: uuid
          description: ID of the wallet to withdraw from
        amount:
          type: number
          format: float
          minimum: 0.01
          description: Amount to withdraw
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the transaction
        type:
          type: string
          enum: [FUND, TRANSFER, WITHDRAWAL]
          description: Type of transaction
        amount:
          type: number
          format: float
          description: Transaction amount
        senderWalletId:
          type: string
          format: uuid
          description: ID of the sender's wallet (null for FUND)
        receiverWalletId:
          type: string
          format: uuid
          description: ID of the receiver's wallet (null for WITHDRAWAL)
        reference:
          type: string
          description: Unique reference for the transaction
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED]
          description: Status of the transaction
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the transaction was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the transaction was last updated
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Error type
paths:
...
  /wallet/fund:
    post:
      summary: Fund a wallet
      description: Add funds to a wallet
      tags:
        - Wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FundRequest'
      responses:
        '201':
          description: Wallet funded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /wallet/transfer:
    post:
      summary: Transfer money between wallets
      description: Transfer money from one wallet to another
      tags:
        - Wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '201':
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid input or insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /wallet/withdraw:
    post:
      summary: Withdraw money from a wallet
      description: Withdraw money from a wallet
      tags:
        - Wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
      responses:
        '201':
          description: Withdrawal successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          description: Invalid input or insufficient funds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Wallet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...